Fixade till klasser och databasen.

+++ hakimlivs/Models/Product.cs
@@ -5,14 +5,19 @@ namespace hakimlivs.Models
     public class Product
     {
         public int Id { get; set; }
+
         [Required]
         public string Name { get; set; }
+
         [Required]
-        public int Price { get; set; }
+        public decimal Price { get; set; }
+
         [Required]
         public string Category { get; set; }
+
         [Required]
         public string Info { get; set; }
+
         [Required]
         public string Image { get; set; }
     }
